Add compatibility on elastic search 2.x
With elastic 2.0 is not allowed anymore to pass the field _id on the document.
So the first thing we notice when try to index firms was the error on
elastic search:

```
MapperParsingException[Field [_id] is a metadata field and cannot be added
inside a document. Use the index API request parameters.]
```

Since we can pass the id on the resource path when you're creating
the documents **one by one** and **not on bulk** then removing
from source and parsing from the root node makes elastic works in
the same way as before.